+++ src/compas/viewers/__old/meshviewer.py
@@ -9,9 +9,9 @@ from compas.geometry import centroid_points
 
 from compas.viewers.viewer import Viewer
 
-from compas.viewers.core.drawing import xdraw_polygons
-from compas.viewers.core.drawing import xdraw_lines
-from compas.viewers.core.drawing import xdraw_points
+from compas.viewers.core.drawing import draw_polygons
+from compas.viewers.core.drawing import draw_lines
+from compas.viewers.core.drawing import draw_points
 
 
 __author__     = 'Tom Van Mele'
@@ -79,10 +79,10 @@ class MeshViewer(Viewer):
         #         'width' : 2.0
         #     })
 
-        xdraw_polygons(polygons)
-        xdraw_lines(lines)
-        xdraw_points(points)
-        # xdraw_lines(normals)
+        draw_polygons(polygons)
+        draw_lines(lines)
+        draw_points(points)
+        # draw_lines(normals)
 
     def keypress(self, key, x, y):
         """