+++ src/compas/viewers/__old/networkviewer.py
@@ -12,8 +12,8 @@ from compas.geometry import centroid_points
 
 from compas.viewers.viewer import Viewer
 
-from compas.viewers.core.drawing import xdraw_points
-from compas.viewers.core.drawing import xdraw_lines
+from compas.viewers.core.drawing import draw_points
+from compas.viewers.core.drawing import draw_lines
 
 from compas.utilities import color_to_colordict
 from compas.utilities import color_to_rgb
@@ -135,9 +135,9 @@ class NetworkViewer(Viewer):
         #                 'color': (0, 1.0, 0),
         #                 'width': 3.0
         #             })
-        xdraw_points(points)
-        xdraw_lines(lines)
-        # xdraw_lines(loads)
+        draw_points(points)
+        draw_lines(lines)
+        # draw_lines(loads)
 
     # --------------------------------------------------------------------------
     # keyboard functionality