+++ src/compas_rhino/artists/artist.py
@@ -173,7 +173,7 @@ class Artist(object):
 
         """
         layer = layer or self.layer
-        return compas_rhino.xdraw_points(points, layer=layer, clear=clear_layer, redraw=redraw)
+        return compas_rhino.draw_points(points, layer=layer, clear=clear_layer, redraw=redraw)
 
     def draw_lines(self, lines, layer=None, clear_layer=False, redraw=False):
         """Draw a collection of lines.
@@ -199,7 +199,7 @@ class Artist(object):
 
         """
         layer = layer or self.layer
-        return compas_rhino.xdraw_lines(lines, layer=layer, clear=clear_layer, redraw=redraw)
+        return compas_rhino.draw_lines(lines, layer=layer, clear=clear_layer, redraw=redraw)
 
     def draw_polylines(self, polylines, layer=None, clear_layer=False, redraw=False):
         """Draw a collection of polygons.
@@ -225,7 +225,7 @@ class Artist(object):
 
         """
         layer = layer or self.layer
-        return compas_rhino.xdraw_polylines(polylines, layer=layer, clear=clear_layer, redraw=redraw)
+        return compas_rhino.draw_polylines(polylines, layer=layer, clear=clear_layer, redraw=redraw)
 
     def draw_polygons(self, polygons, layer=None, clear_layer=False, redraw=False):
         """Draw a collection of polygons.
@@ -251,7 +251,7 @@ class Artist(object):
 
         """
         layer = layer or self.layer
-        return compas_rhino.xdraw_polylines(polygons, layer=layer, clear=clear_layer, redraw=redraw)
+        return compas_rhino.draw_polylines(polygons, layer=layer, clear=clear_layer, redraw=redraw)
 
 
 # ==============================================================================