+++ src/compas_rhino/artists/meshartist.py
@@ -96,7 +96,7 @@ class MeshArtist(FaceArtist, EdgeArtist, VertexArtist, Artist):
                 new_faces.append(face)
         layer = self.layer
         name = "{}.mesh".format(self.mesh.name)
-        return compas_rhino.xdraw_mesh(vertices, new_faces, layer=layer, name=name, color=color)
+        return compas_rhino.draw_mesh(vertices, new_faces, layer=layer, name=name, color=color)
 
     def clear_mesh(self):
         compas_rhino.delete_objects(compas_rhino.get_objects(name="{}.mesh".format(self.mesh.name)))