+++ src/compas_rhino/geometry/surface.py
@@ -449,4 +449,4 @@ if __name__ == '__main__':
             'color' : (0, 255, 0),
         })
 
-    compas_rhino.xdraw_points(points, layer='Layer 01', clear=True, redraw=True)
+    compas_rhino.draw_points(points, layer='Layer 01', clear=True, redraw=True)